Inline the meta box field collection into the save flow
The fields are collected right after each form is submitted, with the
form reached through the submit button, and a document is skipped when
the button is missing because the page is still loading or failed to
load.
