+++ packages/edit-post/src/store/actions.js
@@ -9,10 +9,7 @@ import { addAction } from '@wordpress/hooks';
 import { store as coreStore } from '@wordpress/core-data';
 import { store as noticesStore } from '@wordpress/notices';
 import { __ } from '@wordpress/i18n';
-import {
-	collectMetaBoxFieldsData,
-	getMetaBoxesIframeName,
-} from '../utils/meta-boxes';
+import { getMetaBoxesIframeName } from '../utils/meta-boxes';
 import { unlock } from '../lock-unlock';
 
 const { interfaceStore } = unlock( editorPrivateApis );
@@ -286,31 +283,51 @@ export const requestMetaBoxUpdates =
 		// The frames have to be found through the elements: under the
 		// editor's Document-Isolation-Policy the frames' browsing context
 		// names are cleared, so `window.frames[ name ]` finds nothing.
-		// A document without the post form is still loading or failed to
-		// load, so there is nothing to save in it.
-		const frameDocuments = [ 'main', 'side' ]
+		// A document without the submit button is still loading or failed
+		// to load, so there is nothing to save in it.
+		const submitters = [ 'main', 'side' ]
 			.map( ( location ) =>
 				document.querySelector(
 					`iframe[name="${ getMetaBoxesIframeName( location ) }"]`
 				)
 			)
 			.filter( Boolean )
-			.map( ( iframe ) => iframe.contentDocument )
-			.filter( ( frameDocument ) =>
-				frameDocument?.getElementById( 'post' )
-			);
+			.map( ( iframe ) =>
+				iframe.contentDocument?.getElementById(
+					'gutenberg-meta-box-submitter'
+				)
+			)
+			.filter( Boolean );
 
-		for ( const frameDocument of frameDocuments ) {
+		const metaBoxesFormData = submitters.map( ( submitter ) => {
 			// Some meta boxes, including TinyMCE editors, only write their
 			// values into their form fields in a submit handler, so submit
 			// the form with the hidden button the loader page renders for
 			// this. The loader cancels every submission, so the form does
 			// not actually navigate.
-			const submitter = frameDocument.getElementById(
-				'gutenberg-meta-box-submitter'
+			const { form } = submitter;
+			form.requestSubmit( submitter );
+
+			// The form also holds the title and content fields, which must
+			// not be submitted. Disabled fields are excluded from
+			// `FormData`, so the fields outside the meta boxes are disabled
+			// while the data is constructed.
+			const outsideFields = [ ...form.elements ].filter(
+				( field ) =>
+					! field.disabled &&
+					! field.closest( '.meta-box-sortables .postbox' )
 			);
-			submitter?.form?.requestSubmit( submitter );
-		}
+			for ( const field of outsideFields ) {
+				field.disabled = true;
+			}
+			try {
+				return new form.ownerDocument.defaultView.FormData( form );
+			} finally {
+				for ( const field of outsideFields ) {
+					field.disabled = false;
+				}
+			}
+		} );
 
 		// We gather the base form data.
 		const baseFormData = new window.FormData(
@@ -336,11 +353,7 @@ export const requestMetaBoxUpdates =
 			post.author ? [ 'post_author', post.author ] : false,
 		].filter( Boolean );
 
-		// We gather the meta box fields from the iframes they render in.
-		const formDataToMerge = [
-			baseFormData,
-			...frameDocuments.map( collectMetaBoxFieldsData ),
-		];
+		const formDataToMerge = [ baseFormData, ...metaBoxesFormData ];
 
 		// Merge all form data objects into a single one.
 		const formData = formDataToMerge.reduce( ( memo, currentFormData ) => {
